DracoTown GitHub Preview
This is the GitHub Preview. Not everything works. If you want to start up the project start it in "main.py" via the code software editor.

DracoTown GitHub Preview:
1 - Very Very Early Release of DracoTown
2 - Preview of Inventory / User Info
3 - Preview of Shops
DracoTown GitHub Preview Downsides:
1 - Not All Shops Work
2 - Quests Don't Work
3 - Level Upgrading Does Not Work
4 - Value On Every Item Does Not Work
5 - Eat On Every Edible Item Does Not Work

This Release is only if you are interested in seeing what this project will be about.
The actual Alpha Release will happen when Quests, Levels, Value, and Edible items
work.

From Mashu